refactor(security): relocate writeColumn to safe-writes, sanitize its writes
