+++ __tests__/safe-writes.test.ts
@@ -25,6 +25,7 @@ import {
   writeSafeValueGrid,
   writeSafeRichText,
   writeSafeRichTextGrid,
+  writeColumn,
 } from "../src/server/safe-writes";
 
 describe("sanitizeForCell", () => {
@@ -214,5 +215,50 @@ describe("writeSafeRichTextGrid", () => {
     expect(writtenDangerous.getText()).toBe(
       "[SSI Error: AI response contained an external request formula — output rejected]",
     );
+  });
+});
+
+describe("writeColumn", () => {
+  it("writes values starting at row 2 using a single setValues call", () => {
+    const setValuesMock = jest.fn();
+    const sheet = {
+      getRange: jest.fn().mockReturnValue({ setValues: setValuesMock }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    writeColumn(sheet, 3, ["a", "b", "c"]);
+    expect(sheet.getRange).toHaveBeenCalledWith(2, 3, 3, 1);
+    expect(setValuesMock).toHaveBeenCalledWith([["a"], ["b"], ["c"]]);
+  });
+
+  it("does nothing when values array is empty", () => {
+    const sheet = {
+      getRange: jest.fn(),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    writeColumn(sheet, 1, []);
+    expect(sheet.getRange).not.toHaveBeenCalled();
+  });
+
+  it("applies wrapStrategy to the written range when provided", () => {
+    const setValuesMock = jest.fn();
+    const setWrapStrategyMock = jest.fn();
+    const sheet = {
+      getRange: jest
+        .fn()
+        .mockReturnValue({ setValues: setValuesMock, setWrapStrategy: setWrapStrategyMock }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    const wrapStrategy = "CLIP" as unknown as GoogleAppsScript.Spreadsheet.WrapStrategy;
+    writeColumn(sheet, 3, ["a", "b"], wrapStrategy);
+    expect(setWrapStrategyMock).toHaveBeenCalledWith(wrapStrategy);
+  });
+
+  it("sanitizes a dangerous value before writing", () => {
+    const setValuesMock = jest.fn();
+    const sheet = {
+      getRange: jest.fn().mockReturnValue({ setValues: setValuesMock }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    writeColumn(sheet, 3, ['=IMAGE("evil.com")', "safe"]);
+    expect(setValuesMock).toHaveBeenCalledWith([
+      ["[SSI Error: AI response contained an external request formula — output rejected]"],
+      ["safe"],
+    ]);
   });
 });
